fix: fix types

+++ common/models.py
@@ -75,7 +75,7 @@ class KillScore(Player):
     rank: int
     kills: dict[str, int] = field(default_factory=dict)
     achievements: dict[str, int] = field(default_factory=dict)
-    ratio: int | None = field(init=False)
+    ratio: float | None = field(init=False)
 
     def __post_init__(self):
         self.ratio = (
